requirements_venv.txt
# ./requirements.txt

# main libs
#discord.py==1.7.3
#web.py==0.62
#mysqlclient
#debugpy

# pip freeze:

aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==23.1.0
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
cheroot==10.0.0
debugpy==1.8.0
discord.py==1.7.3
idna==3.6
jaraco.functools==4.0.0
#mod-wsgi==5.0.0
more-itertools==10.1.0
multidict==6.0.4
#mysqlclient==2.2.0
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
typing_extensions==4.8.0
urllib3==2.1.0
web.py==0.62
wsgilog==0.3.1
yarl==1.9.3